requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0.13
django-crispy-forms==1.9.2
django-import-export==2.4.0
django-map-widgets==0.1.9
django-utils-six==2.0
Django-Verify-Email==0.0.3
django-widget-tweaks==1.4.8
djangorestframework==3.8.2
et-xmlfile==1.0.1
googlemaps==4.4.2
idna==2.10
install==1.3.4
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.1
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
protobuf==3.14.0
psycopg2==2.8.6
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
